requirements.txt
PySide6
qt-material
PyOpenGL
numpy
Pillow
